refactor(results): consolidate HTTP compatibility support
